+++ typescript/crosschain/query_pb.d.ts
@@ -5,7 +5,6 @@
 
 import type { BinaryReadOptions, FieldList, JsonReadOptions, JsonValue, PartialMessage, PlainMessage } from "@bufbuild/protobuf";
 import { Message, proto3 } from "@bufbuild/protobuf";
-import type { Params } from "./params_pb.js";
 import type { OutTxTracker } from "./out_tx_tracker_pb.js";
 import type { PageRequest, PageResponse } from "../cosmos/base/query/v1beta1/pagination_pb.js";
 import type { InTxTracker } from "./in_tx_tracker_pb.js";
@@ -55,55 +54,6 @@ export declare class QueryZetaAccountingResponse extends Message<QueryZetaAccoun
   static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): QueryZetaAccountingResponse;
 
   static equals(a: QueryZetaAccountingResponse | PlainMessage<QueryZetaAccountingResponse> | undefined, b: QueryZetaAccountingResponse | PlainMessage<QueryZetaAccountingResponse> | undefined): boolean;
-}
-
-/**
- * QueryParamsRequest is request type for the Query/Params RPC method.
- *
- * @generated from message zetachain.zetacore.crosschain.QueryParamsRequest
- */
-export declare class QueryParamsRequest extends Message<QueryParamsRequest> {
-  constructor(data?: PartialMessage<QueryParamsRequest>);
-
-  static readonly runtime: typeof proto3;
-  static readonly typeName = "zetachain.zetacore.crosschain.QueryParamsRequest";
-  static readonly fields: FieldList;
-
-  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): QueryParamsRequest;
-
-  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): QueryParamsRequest;
-
-  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): QueryParamsRequest;
-
-  static equals(a: QueryParamsRequest | PlainMessage<QueryParamsRequest> | undefined, b: QueryParamsRequest | PlainMessage<QueryParamsRequest> | undefined): boolean;
-}
-
-/**
- * QueryParamsResponse is response type for the Query/Params RPC method.
- *
- * @generated from message zetachain.zetacore.crosschain.QueryParamsResponse
- */
-export declare class QueryParamsResponse extends Message<QueryParamsResponse> {
-  /**
-   * params holds all the parameters of this module.
-   *
-   * @generated from field: zetachain.zetacore.crosschain.Params params = 1;
-   */
-  params?: Params;
-
-  constructor(data?: PartialMessage<QueryParamsResponse>);
-
-  static readonly runtime: typeof proto3;
-  static readonly typeName = "zetachain.zetacore.crosschain.QueryParamsResponse";
-  static readonly fields: FieldList;
-
-  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): QueryParamsResponse;
-
-  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): QueryParamsResponse;
-
-  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): QueryParamsResponse;
-
-  static equals(a: QueryParamsResponse | PlainMessage<QueryParamsResponse> | undefined, b: QueryParamsResponse | PlainMessage<QueryParamsResponse> | undefined): boolean;
 }
 
 /**