+++ x/crosschain/keeper/grpc_query_last_block_height_test.go
@@ -1,6 +1,7 @@
 package keeper
 
 import (
+	"math"
 	"testing"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
@@ -51,6 +52,39 @@ func TestLastBlockHeightQuerySingle(t *testing.T) {
 			}
 		})
 	}
+}
+
+func TestLastBlockHeightLimits(t *testing.T) {
+	t.Run("should err if last send height is max int", func(t *testing.T) {
+		keeper, ctx := setupKeeper(t)
+		wctx := sdk.WrapSDKContext(ctx)
+		keeper.SetLastBlockHeight(ctx, types.LastBlockHeight{
+			Index:          "index",
+			LastSendHeight: math.MaxInt64,
+		})
+
+		res, err := keeper.LastBlockHeight(wctx, &types.QueryGetLastBlockHeightRequest{
+			Index: "index",
+		})
+		require.Nil(t, res)
+		require.Error(t, err)
+	})
+
+	t.Run("should err if last receive height is max int", func(t *testing.T) {
+		keeper, ctx := setupKeeper(t)
+		wctx := sdk.WrapSDKContext(ctx)
+		keeper.SetLastBlockHeight(ctx, types.LastBlockHeight{
+			Index:             "index",
+			LastSendHeight:    10,
+			LastReceiveHeight: math.MaxInt64,
+		})
+
+		res, err := keeper.LastBlockHeight(wctx, &types.QueryGetLastBlockHeightRequest{
+			Index: "index",
+		})
+		require.Nil(t, res)
+		require.Error(t, err)
+	})
 }
 
 func TestLastBlockHeightQueryPaginated(t *testing.T) {