+++ x/crosschain/module_simulation.go
@@ -15,9 +15,7 @@ func (AppModule) GenerateGenesisState(simState *module.SimulationState) {
 	for i, acc := range simState.Accounts {
 		accs[i] = acc.Address.String()
 	}
-	crosschainGenesis := types.GenesisState{
-		Params: types.DefaultParams(),
-	}
+	crosschainGenesis := types.GenesisState{}
 	simState.GenState[types.ModuleName] = simState.Cdc.MustMarshalJSON(&crosschainGenesis)
 }